extract GeneratePackageAnnotationsTask from AndroidNonNullPlugin
